LayoutTests/inspector/unit-tests/protocol-test-dispatch-event-to-frontend-expected.txt


== Running test suite: ProtocolTest.TestPage.dispatchEventToFrontend
-- Running test case: ProtocolTest.TestPage.dispatchEventToFrontend
PASS: Received Event
PASS: Event name should be the event dispatched.
PASS: Event data should be included.
